Another bad hack for BJ backward compatibility.
This *really* is a bad hack, as it abuses an non-public side-effect of the
attributes getter and setter callbacks.  I am quite ashamed of this code :-/

On the long run, BJ should be cleaned up to use dicts, or to at least accept
env dicts when querying a job description, and this hack should be reverted.

A.