remove timout garbage collector
This is not a viable approach -- the initialize/finalize hooks create reference cycles which screw python garbage collection.  One could use weakrefs, but if we cannot guarantee that a connection which got timed out can be re-instantiated, then we should not mess with the connection in the first place.

However, this commit still does not solve the garbage collection problem -- while the adaptor destructors are now correctly called at the end of the program execution, I see no sign of any earlier garbage collection activities, no matter the runtime, obj usage etc.  So, this needs further debugging...