concurrency for benchmarks
This is somewhat crudely hacked together, apologies, but the benchmark suite can now use similar config files as the test suite (the same in fact, but with an additional section specifying the scope of the benchmarks), and the benchmarks now support concurrency, i.e. the can distribute the load to measure over multiple threads.