added WALL_TIME_LIMIT to all adaptors
unfortunately, BigJob sets this as a default value. Adding it in the PBS case actually makes sense. No idea why it wasn't there in the first place because the adaptor actually truies to evalute jd.wall_time_limit. In the 'shell' adaptor case it isn't really the best idea, but unfortunately inevitable.