add copied_files metric to copy tasks.
This commit also adjusts a number of things in saga.Task to allow for arbitrary metrics to be added.  Those changes need to be reflected in *all* @SYNC calls which are supposed to be used in async tasks -- they now get an additional parameter "_from_task", which, you guess it, contains the task object itself, so that metrics etc. can be registered and updated.